We are first to define several default functions storages for the overall class setting of the simulations. Some lists and dict are created in order to store and for the future handle of problems.
The 1st function: Parses a line starting with 'DEVICE' and updates device information.
The 2nd function: Parses a line starting with 'PROPAGATE' and stores propagation rules
The 3rd function: Parses a line starting with 'ALERT' and stores alert
The 4th: Parses a line starting with 'ALERT' and stores alerts
The 5th: Parses a line starting with 'CANCEL' and stores cancellations
The 6th: Reads and parses the input file
The 7th: The overall simulation logical management
The 8th: Reads the input file path from the standard input and checks for existence
The 9th and the last: runs the simulation.